finished sla-calculation methods
